+++ src/ui-store.c
@@ -302,9 +302,9 @@ static void store_display_entry(struct menu *menu, int oid, bool cursor, int row
 
 		/* Actually draw the price */
 		if (tval_can_have_charges(obj) && (obj->number > 1))
-			strnfmt(out_val, sizeof out_val, "%9ld avg", (long)x);
+			strnfmt(out_val, sizeof out_val, "%9d avg", x);
 		else
-			strnfmt(out_val, sizeof out_val, "%9ld    ", (long)x);
+			strnfmt(out_val, sizeof out_val, "%9d    ", x);
 
 		c_put_str(colour, out_val, row, ctx->scr_places_x[LOC_PRICE]);
 	}
@@ -343,8 +343,8 @@ static void store_display_frame(struct store_context *ctx)
 		put_str(owner_name, ctx->scr_places_y[LOC_OWNER], 1);
 
 		/* Show the max price in the store (above prices) */
-		strnfmt(buf, sizeof(buf), "%s (%ld)", store_name,
-				(long)(proprietor->max_cost));
+		strnfmt(buf, sizeof(buf), "%s (%d)", store_name,
+				proprietor->max_cost);
 		prt(buf, ctx->scr_places_y[LOC_OWNER],
 			ctx->scr_places_x[LOC_OWNER] - strlen(buf));
 
@@ -435,7 +435,7 @@ static void store_redraw(struct store_context *ctx)
 	}
 
 	if (ctx->flags & (STORE_GOLD_CHANGE)) {
-		prt(format("Gold Remaining: %9ld", (long)player->au),
+		prt(format("Gold Remaining: %9d", player->au),
 				ctx->scr_places_y[LOC_AU], ctx->scr_places_x[LOC_AU]);
 		ctx->flags &= ~(STORE_GOLD_CHANGE);
 	}
